+++ src/bittytax/conv/parsers/__init__.py
@@ -22,6 +22,7 @@ from . import (
     coinlist,
     coinmetro,
     coinomi,
+    cointracker,
     cointracking,
     cryptocom,
     cryptopia,